npm-agent: use setitimer instead of timerfd
setitimer + self-pipe is more portable. Since this interrupts the poll,
we also now check errno to see if poll failed due to an interrupt. If
it did, then we just restart the loop and continue as normal.
Consequently, the running boolean actually does something now, instead
of npm-agent stopping simply because poll failed.